Fix cannot find module problem
When running before, got an error:
Error: Cannot find module 'webpack-cli/bin/config-yargs'

Following this Stack Overflow post:
https://stackoverflow.com/questions/40379139/cannot-find-module-webpack-bin-config-yargs

The main change is to the `start` script in `package.json` to use
`webpack serve` in place of `webpack-dev-server`.

Also check in the update to `package-lock.json` that happened when I ran it.

This change makes the sample work as documented on my machine.